requirements.txt
flask
pandas
numpy
onnxruntime
joblib
sqlalchemy
pymysql
psycopg2-binary
cryptography
gunicorn
python-dotenv
